// src/app/models/wireconsumption.ts
export class Wireconsumption {

    Plant?: string;
    Supplier?: string;
    Week_Number?: number;
    yearB?: number;
    Initial_Stock?: number;
    Quantity_received?: number;
    Real_Consumption?: number;
    Final_Stock?: number;
    MaterialBreaks?: number;
    ProcessBreaks?: number;
    Rate?: number;
}